finished applying point mechanism

+++ Jelwelry Sale System At The Store/src/main/java/com/swp391/jewelrysalesystem/models/Customer.java
@@ -11,5 +11,5 @@ public class Customer {
     private String name;
     private String contactInfo;
     private String gender;
-    private double point;
+    private int point;
 }
